Partial fix #12
Fix the parameter save fp mode is always 0 for the StoreBarrier calls.
This parameter is tuned by the live outs set of a call instruction, if
the live outs set of a call instruction contains any registers are not
one of the general purpose registers, then save fp mode will be set to
kSaveFPMode.