Fix #12.
When a function's calling convention is V8SBCC, LLVM will check a call site:
1. If it's a C call.
2. If it has a function attribute "safe-fp".

If the conditions are match, the fp registers that could be clobbered by
a C function will be save to a stack area.

The v8 side will set the "safe-fp" attribute to a store barrier builtin
function if the descriptor's save fp mode is kSafeFP.
There is no need to handle the descriptor's save fp mode other than a
store battier builtin function. LLVM will calculate this constraint
properly.