Fix some val fails to be rematerialized bug.
The val split by calls may not get rematerialized, for it's the VNInfos of its
live range are split by phis. When the LiveRangeEdit::scanRemattable
tried to scan value for rematerialize, which essentially check the
VNInfo of a reg before any split to see if its def fit for
rematerialize, it will get null def MI at the end. Because it's a phi
def before any split.

The new patch make the single def virtual registers' valnos combine into
one, if they are all phi-def or unused except one def.